+++ src/main/scala/com/easyforger/creatures/EFCustomZombie.scala
@@ -7,7 +7,7 @@ package com.easyforger.creatures
 import net.minecraft.entity.monster.EntityZombie
 import net.minecraft.world.World
 
-case class ZombieConfig(common: CommonEntityConfig = CommonEntityConfig()) extends CreatureConfig
+case class ZombieConfig(common: CommonEntityConfig = CommonEntityConfig(None, None, None)) extends CreatureConfig
 
 class EFCustomZombie(world: World) extends EntityZombie(world) with CommonCustomMonster {
   val zombie = VanillaCreatures.zombieConfig