+++ Changelog.md
@@ -4,6 +4,7 @@
 * Updated forge to 1.10.2-12.18.3.2185;
 * Changed `EFItemFood.addPotionEffect` to take a `PotionEffect` instead of individual parameters;
 * Renamed `CustomCreeper`, `CustomSkeleton` and `CustomZombie` to `EFCustomCreeper`, `EFCustomSkeleton` and `EFCustomZombie`.
+* Replaced `common.heldItem` with `common.heldItemMainHand` and `common.heldItemOffHand`
 
 ===================================== Version 0.5 =====================================